Test e2e y test unitarios

+++ src/controller/users.js
@@ -1,62 +1,71 @@
 const {
   createUserAndRole,
-  getUsers,
+  getUsersWithPagination,
   getUserByIdOrEmail,
   getUserByEmail,
   updateSingle,
   deleteSingle,
-} = require("../services/users");
+} = require('../services/users');
+const { isAdmin } = require('../middleware/auth');
+
+const { isValidEmail, convertToLinks, isValidPassword } = require('../utils/util');
 
 module.exports = {
   getUsers: async (req, res) => {
     const { page = 1, limit = 10 } = req.query;
-    const host = req.headers.host; // localhost:80
-    const totalPages = Math.ceil((await getUsers(1, 0)).length / limit);
-    const users = await getUsers(Number(page), Number(limit));
-    const link = {
-      first: page === 1 ? "" : `http://${host}/users?page=1`,
-      next: page === totalPages ? "" : `http://${host}/users?page=${Number(page) + 1}`,
-      prev: page === 1 ? "" : `http://${host}/users?page=${Number(page) - 1}`,
-      last: page === totalPages ? "" : `http://${host}/users?page=${totalPages}`,
-    };
-    res.set("Link", JSON.stringify(link));
-    res.json(users);
+    const { protocol, host } = req.headers;
+    const totalPages = Math.ceil((await getUsersWithPagination(1, 0)).length / limit);
+    const users = await getUsersWithPagination(Number(page), Number(limit));
+    const link = convertToLinks(protocol, host, 'users', limit, page, totalPages);
+    res.set('links', JSON.stringify(link));
+    return res.json(users);
   },
-  getUserById: async (req, res) => {
+  getUserById: async (req, res, next) => {
     const user = await getUserByIdOrEmail(req.params.uid);
     if (!user) return res.status(404).json({ message: `User: ${req.params.uid} does not exists` });
-    res.json(user);
+    if (req.userToken.id === user._id.toString() || await isAdmin(req)) {
+      return res.json(user);
+    }
+    return next(403);
   },
   createUser: async (req, res) => {
     const { email, password, roles } = req.body;
-    if (!email || !password) return res.status(400).json({ message: "Email or password not found" });
+    if (!email || !password) return res.status(400).json({ message: 'Email or password is missing' });
+    if (!isValidEmail(email) || !isValidPassword(password)) return res.status(400).json({ message: 'Incorrect format of email or password' });
     const userFound = await getUserByEmail(email);
     if (userFound) return res.status(403).json({ message: `User with email: ${email} already exists` });
     const user = await createUserAndRole(email, password, roles);
-    res.json(user);
+    return res.json(user);
   },
   putUser: async (req, res) => {
+    const admin = await isAdmin(req);
     const user = await getUserByIdOrEmail(req.params.uid);
     if (!user) return res.status(404).json({ message: `User: ${req.params.uid} does not exists` });
-    const admin = req.userToken.admin;
-    // if (req.userToken.id !== user._id.toString() && !admin) {
-    //   return res.status(403).json({ message: "Admin permission is required or be the same user that wants modified" });
-    // }
-    const { email, password, roles } = req.body;
-    if (!email || !password) return res.status(400).json({ message: "Email or password not found" });
-    if (roles && !admin) {
-      return res.status(403).json({ message: "Admin permission is required to modified roles value" });
+    if (req.userToken.id !== user._id.toString() && !admin) {
+      return res.status(403).json({ message: 'Admin permission is required or be the owner' });
     }
+    let { email, password, roles } = req.body;
+    if (roles && !admin) {
+      return res.status(403).json({ message: 'Admin permission is required to modified roles value' });
+    }
+    if ((Object.keys(req.body).length === 0) || email === '' || password === '') {
+      return res.status(400).json({ message: 'Email or password is missing' });
+    }
+    if (email && !isValidEmail(email)) return res.status(400).json({ message: 'Incorrect format of email' });
+    if (password && !isValidPassword(password)) return res.status(400).json({ message: 'Incorrect format of password' });
+    if (!password) (password = user.password);
+    if (!email) (email = user.email);
+    if (!roles) (roles = user.roles);
     const userUpdate = await updateSingle(user._id, user.roles._id, email, password, roles);
-    res.json(userUpdate);
+    return res.json(userUpdate);
   },
   deleteUser: async (req, res) => {
     const user = await getUserByIdOrEmail(req.params.uid);
     if (!user) return res.status(404).json({ message: `User: ${req.params.uid} does not exists` });
-    // if (req.userToken.id !== user._id.toString() && !req.userToken.admin) {
-    //   return res.status(403).json({ message: "Admin permission is required or be the same user that wants delete" });
-    // }
+    if (req.userToken.id !== user._id.toString() && !await isAdmin(req)) {
+      return res.status(403).json({ message: 'Admin permission is required or be the owner' });
+    }
     await deleteSingle(user._id, user.roles._id);
-    res.json(user);
+    return res.json(user);
   },
 };
